cart, favourite,order controller, service created

+++ src/main/java/com/vipusa/onlineFood/repository/CartRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartRepository extends CrudRepository<Cart, Integer> {
+public interface CartRepository extends CrudRepository<Cart, Long> {
 }
